+++ app/src/androidTest/java/de/test/antennapod/ui/PreferencesTest.java
@@ -173,6 +173,7 @@ public class PreferencesTest {
     @Test
     public void testAutoDelete() {
         clickPreference(R.string.downloads_pref);
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
         final boolean autoDelete = UserPreferences.isAutoDelete();
         onView(withText(R.string.pref_auto_delete_title)).perform(click());
         Awaitility.await().atMost(1000, MILLISECONDS)
@@ -185,8 +186,10 @@ public class PreferencesTest {
     @Test
     public void testAutoDeleteLocal() {
         clickPreference(R.string.downloads_pref);
-        final boolean initialAutoDelete = UserPreferences.isAutoDeleteLocal();
-        assertFalse(initialAutoDelete);
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
+        assertTrue(UserPreferences.isAutoDelete());
+        assertFalse(UserPreferences.isAutoDeleteLocal());
 
         onView(withText(R.string.pref_auto_local_delete_title)).perform(click());
         onView(withText(R.string.yes)).perform(click());
@@ -289,7 +292,7 @@ public class PreferencesTest {
     @Test
     public void testEpisodeCleanupFavoriteOnly() {
         clickPreference(R.string.downloads_pref);
-        onView(withText(R.string.pref_automatic_download_title)).perform(click());
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
         onView(withText(R.string.pref_episode_cleanup_title)).perform(click());
         onView(withId(R.id.select_dialog_listview)).perform(swipeDown());
         onView(withText(R.string.episode_cleanup_except_favorite_removal)).perform(click());
@@ -300,7 +303,7 @@ public class PreferencesTest {
     @Test
     public void testEpisodeCleanupQueueOnly() {
         clickPreference(R.string.downloads_pref);
-        onView(withText(R.string.pref_automatic_download_title)).perform(click());
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
         onView(withText(R.string.pref_episode_cleanup_title)).perform(click());
         onView(withId(R.id.select_dialog_listview)).perform(swipeDown());
         onView(withText(R.string.episode_cleanup_queue_removal)).perform(click());
@@ -311,7 +314,7 @@ public class PreferencesTest {
     @Test
     public void testEpisodeCleanupNeverAlg() {
         clickPreference(R.string.downloads_pref);
-        onView(withText(R.string.pref_automatic_download_title)).perform(click());
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
         onView(withText(R.string.pref_episode_cleanup_title)).perform(click());
         onView(withId(R.id.select_dialog_listview)).perform(swipeUp());
         onView(withText(R.string.episode_cleanup_never)).perform(click());
@@ -322,7 +325,7 @@ public class PreferencesTest {
     @Test
     public void testEpisodeCleanupClassic() {
         clickPreference(R.string.downloads_pref);
-        onView(withText(R.string.pref_automatic_download_title)).perform(click());
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
         onView(withText(R.string.pref_episode_cleanup_title)).perform(click());
         onView(withText(R.string.episode_cleanup_after_listening)).perform(click());
         Awaitility.await().atMost(1000, MILLISECONDS)
@@ -339,7 +342,7 @@ public class PreferencesTest {
     @Test
     public void testEpisodeCleanupNumDays() {
         clickPreference(R.string.downloads_pref);
-        clickPreference(R.string.pref_automatic_download_title);
+        onView(withText(R.string.pref_auto_delete_title)).perform(click());
         clickPreference(R.string.pref_episode_cleanup_title);
         String search = res.getQuantityString(R.plurals.episode_cleanup_days_after_listening, 3, 3);
         onView(withText(search)).perform(scrollTo());
