+++ ui/preferences/src/main/res/xml/preferences_auto_deletion.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen
+        xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:enabled="true"
+        android:key="prefAutoDelete"
+        android:summary="@string/pref_auto_delete_sum"
+        android:title="@string/pref_auto_delete_title"/>
+    <SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:enabled="true"
+        android:key="prefAutoDeleteLocal"
+        android:summary="@string/pref_auto_local_delete_sum"
+        android:title="@string/pref_auto_local_delete_title"/>
+    <SwitchPreferenceCompat
+        android:defaultValue="true"
+        android:enabled="true"
+        android:key="prefFavoriteKeepsEpisode"
+        android:summary="@string/pref_favorite_keeps_episodes_sum"
+        android:title="@string/pref_favorite_keeps_episodes_title"/>
+    <de.danoeh.antennapod.ui.preferences.preference.MaterialListPreference
+        android:defaultValue="-1"
+        android:entries="@array/episode_cleanup_entries"
+        android:key="prefEpisodeCleanup"
+        android:title="@string/pref_episode_cleanup_title"
+        android:summary="@string/pref_episode_cleanup_summary"
+        android:entryValues="@array/episode_cleanup_values"/>
+
+</PreferenceScreen>
