+++ ui/preferences/src/main/res/xml/preferences_autodownload.xml
@@ -15,13 +15,6 @@
             android:title="@string/pref_episode_cache_title"
             android:summary="@string/pref_episode_cache_summary"
             android:entryValues="@array/episode_cache_size_values"/>
-    <de.danoeh.antennapod.ui.preferences.preference.MaterialListPreference
-            android:defaultValue="-1"
-            android:entries="@array/episode_cleanup_entries"
-            android:key="prefEpisodeCleanup"
-            android:title="@string/pref_episode_cleanup_title"
-            android:summary="@string/pref_episode_cleanup_summary"
-            android:entryValues="@array/episode_cleanup_values"/>
     <SwitchPreferenceCompat
             android:key="prefEnableAutoDownloadOnBattery"
             android:title="@string/pref_automatic_download_on_battery_title"
