+++ ui/preferences/src/main/res/xml/preferences_downloads.xml
@@ -27,24 +27,10 @@
             android:key="prefAutoDownloadSettings"
             android:title="@string/pref_automatic_download_title"
             search:ignore="true" />
-        <SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:enabled="true"
-            android:key="prefAutoDelete"
+        <Preference
+            android:key="prefAutoDeleteScreen"
             android:summary="@string/pref_auto_delete_sum"
             android:title="@string/pref_auto_delete_title"/>
-        <SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:enabled="true"
-            android:key="prefAutoDeleteLocal"
-            android:summary="@string/pref_auto_local_delete_sum"
-            android:title="@string/pref_auto_local_delete_title"/>
-        <SwitchPreferenceCompat
-            android:defaultValue="true"
-            android:enabled="true"
-            android:key="prefFavoriteKeepsEpisode"
-            android:summary="@string/pref_favorite_keeps_episodes_sum"
-            android:title="@string/pref_favorite_keeps_episodes_title"/>
         <SwitchPreferenceCompat
             android:defaultValue="false"
             android:enabled="true"
